Extract HmdaActor to persistence-model
