+++ persistence/src/test/scala/hmda/persistence/institutions/InstitutionPersistenceSpec.scala
@@ -1,7 +1,6 @@
 package hmda.persistence.institutions
 
 import akka.testkit.{ EventFilter, TestProbe }
-import hmda.actor.test.ActorSpec
 import hmda.model.institution.Agency.{ CFPB, FDIC }
 import hmda.model.institution.ExternalIdType.{ FdicCertNo, FederalTaxId, RssdId }
 import hmda.model.institution.{ ExternalId, Institution }
@@ -10,6 +9,7 @@ import hmda.model.institution.InstitutionType.Bank
 import hmda.persistence.messages.CommonMessages.GetState
 import hmda.persistence.demo.DemoData
 import hmda.persistence.institutions.InstitutionPersistence.{ CreateInstitution, GetInstitutionById, ModifyInstitution, _ }
+import hmda.persistence.model.ActorSpec
 
 class InstitutionPersistenceSpec extends ActorSpec {
