+++ persistence/src/test/scala/hmda/persistence/processing/HmdaFileParserSpec.scala
@@ -6,10 +6,10 @@ import akka.actor.{ ActorRef, ActorSystem }
 import akka.testkit.{ EventFilter, TestProbe }
 import org.scalatest.BeforeAndAfterEach
 import com.typesafe.config.ConfigFactory
-import hmda.actor.test.ActorSpec
 import hmda.model.fi.SubmissionId
 import hmda.parser.fi.ts.TsCsvParser
 import hmda.persistence.messages.CommonMessages.GetState
+import hmda.persistence.model.ActorSpec
 import hmda.persistence.processing.HmdaFileParser._
 import hmda.persistence.processing.HmdaRawFile._
