+++ persistence/src/test/scala/hmda/persistence/processing/HmdaRawFileSpec.scala
@@ -4,9 +4,9 @@ import java.time.Instant
 
 import akka.testkit.TestProbe
 import com.typesafe.config.ConfigFactory
-import hmda.actor.test.ActorSpec
 import hmda.model.fi.SubmissionId
 import hmda.persistence.messages.CommonMessages._
+import hmda.persistence.model.ActorSpec
 import hmda.persistence.processing.HmdaRawFile._
 
 class HmdaRawFileSpec extends ActorSpec {
