+++ persistence/src/test/scala/hmda/persistence/processing/LocalHmdaEventProcessorSpec.scala
@@ -5,12 +5,12 @@ import akka.pattern.ask
 import akka.testkit.{ EventFilter, TestProbe }
 import akka.util.Timeout
 import com.typesafe.config.ConfigFactory
-import hmda.actor.test.ActorSpec
 import hmda.model.fi._
 import hmda.persistence.messages.CommonMessages.{ Event, GetState }
 import hmda.persistence.HmdaSupervisor._
 import hmda.persistence.institutions.SubmissionPersistence
 import hmda.persistence.institutions.SubmissionPersistence._
+import hmda.persistence.model.ActorSpec
 import hmda.persistence.processing.HmdaFileParser.{ ParsingCompleted, ParsingCompletedWithErrors, ParsingStarted }
 import hmda.persistence.processing.HmdaFileValidator.{ ValidationCompleted, ValidationCompletedWithErrors, ValidationStarted }
 import hmda.persistence.processing.HmdaRawFile.{ UploadCompleted, UploadStarted }
